refactor(otel): load the instrumentor lazily so rxconfig imports stay cheap
`from reflex_otel import OtelPlugin` in rxconfig.py runs in every CLI
process (and the config is loaded several times per run), but the package
import also pulled in opentelemetry.instrumentation and wrapt for the
instrumentor, about 30 ms per load that only the backend needs. Move
ReflexInstrumentor to reflex_otel.instrumentor and resolve it from the
package on first use; `from reflex_otel import ReflexInstrumentor` and the
opentelemetry_instrumentor entry point keep working.

Rebase note: this moves the body of reflex_otel/__init__.py, which the
PR 1/3 fixes also change; when rebasing, carry those changes into
instrumentor.py.
